feat: separate types into own files
