+++ clients/design-system/src/components/Alert/AlertSeverity.type.ts
@@ -1,0 +1,6 @@
+export enum AlertSeverity {
+  INFO = 'info',
+  SUCCESS = 'success',
+  WARNING = 'warning',
+  DANGER = 'danger',
+}
