+++ clients/design-system/src/components/Card/CardRibbonColor.type.ts
@@ -1,0 +1,5 @@
+export enum CardRibbonColor {
+  PRIMARY = 'primary',
+  SECONDARY = 'secondary',
+  TERIARY = 'tertiary',
+}
