+++ clients/design-system/src/components/index.ts
@@ -13,6 +13,7 @@ export * from './SocialMediaBar/SocialMediaBar';
 export * from './Link/NonInteractiveLink';
 export * from './Typography/Typography';
 export * from './Alert/Alert';
+export * from './Alert/AlertSeverity.type';
 export * from './TextField/TextField';
 export * from './Checkbox/Checkbox';
 export * from './Spinner/Spinner';
@@ -20,6 +21,7 @@ export * from './Button/Button';
 export * from './Link/Link';
 export * from './Layout/Layout';
 export * from './Card/BasicCard';
+export * from './Card/CardRibbonColor.type';
 export * from './Paper/Paper';
 export * from './Hero/Hero';
 export * from './Images/Images';
